refactor(AW-23): move gradebook consts out of src/consts, trim comments
