+++ src/pages/TeacherSubjects/components/FinalGradesDetails/StudentCourseGrades.tsx
@@ -6,19 +6,9 @@ import { FormattedMessage } from 'umi';
 import type { StudentGradeRow } from './types';
 import { buildGradeRows, getGradeDisplay, mergeExpandedKeys } from './utils';
 
-/**
- * AW-23 — "Final grades" grades table for a single student.
- *
- * A single tree table: one parent row per course (name + completion tag), whose
- * children are that course's flagged quiz/project results. Data comes from
- * GET .../groups/{group_id}/users/{student_id}/courses-grades (counts_to_grade only).
- */
-
 const formatScore = (score?: number | null, maxScore?: number | null): string =>
   score === null || score === undefined ? '-' : `${score} / ${maxScore ?? '-'}`;
 
-// Pass/fail is co-located with the grade. Renders nothing until the backend computes it
-// (is_passed is null for projects and currently for quizzes too).
 const PassFailTag: React.FC<{ passed?: boolean | null }> = ({ passed }) => {
   if (passed === null || passed === undefined) {
     return null;
@@ -34,20 +24,10 @@ const PassFailTag: React.FC<{ passed?: boolean | null }> = ({ passed }) => {
   );
 };
 
-// Tint for the grade column, the point of the table. Opaque rather than translucent: the
-// column is `fixed: 'right'`, so it overlays the cells scrolling beneath it and any
-// transparency would let them bleed through. These are the composited equivalents of 4% /
-// 8% black over the white table background.
 const GRADE_CELL_BG = '#f5f5f5';
 const GRADE_HEADER_BG = '#ebebeb';
-
-// Below this the three fixed columns (3 × 180) would squeeze the flexible name column past
-// readability, so the table scrolls horizontally instead of compressing further.
 const MIN_TABLE_WIDTH = 986;
 
-// The last column: whichever value we have leads, rendered prominently — the grade when the
-// backend sends one, otherwise the percentage (the `grade` field is not live yet). The
-// percentage is only demoted to muted context when there is a grade to outrank it.
 const GradeCell: React.FC<{ row: StudentGradeRow }> = ({ row }) => {
   const { grade, percent } = getGradeDisplay(row.grade, row.result_percent);
   const primary = grade ?? percent;
@@ -130,9 +110,6 @@ export const StudentCourseGrades: React.FC<Props> = ({ data, loading }) => {
   const rows = useMemo(() => buildGradeRows(data ?? []), [data]);
   const [expandedKeys, setExpandedKeys] = useState<React.Key[]>([]);
   const seenCourseKeysRef = useRef<Set<string>>(new Set());
-
-  // Expand every course on first load, then preserve the user's expand/collapse choices
-  // across data changes (only newly-appearing courses auto-expand).
   useEffect(() => {
     const courseKeys = rows.map((row) => row.key);
     const seen = seenCourseKeysRef.current;
